refactor (ontology cache): move IResourceClassAndPropertyDefinitions to separate file
