+++ src/index.ts
@@ -80,10 +80,10 @@ export { ApiResponseError } from "./models/api-response-error";
 export { ReadOntology } from "./models/v2/ontologies/read-ontology";
 export { OntologyMetadata, OntologiesMetadata } from "./models/v2/ontologies/ontology-metadata";
 export {
-    IResourceClassAndPropertyDefinitions,
     ResourceClassDefinitionWithPropertyDefinition,
     IHasPropertyWithPropertyDefinition
 } from "./cache/ontology-cache/resource-class-definition-with-property-definition";
+export { IResourceClassAndPropertyDefinitions } from "./cache/ontology-cache/resource-class-and-property-definitions";
 export { ClassDefinition, IHasProperty, Cardinality } from "./models/v2/ontologies/class-definition";
 export { SystemPropertyDefinition } from "./models/v2/ontologies/system-property-definition";
 export { ResourcePropertyDefinition } from "./models/v2/ontologies/resource-property-definition";
