+++ src/models/v2/resources/read/read-resource.ts
@@ -1,5 +1,4 @@
 import { JsonObject, JsonProperty } from "json2typescript";
-import { IResourceClassAndPropertyDefinitions } from "../../../../cache/ontology-cache/resource-class-definition-with-property-definition";
 import { Constants } from "../../Constants";
 import { DateTimeStampConverter } from "../../custom-converters/date-time-stamp-converter";
 import { IdConverter } from "../../custom-converters/id-converter";
@@ -9,6 +8,7 @@ import { ResourcePropertyDefinition } from "../../ontologies/resource-property-d
 import { ReadWriteResource } from "../read-write-resource";
 import { TypeGuard } from "../type-guard";
 import { ReadValue } from "../values/read/read-value";
+import { IResourceClassAndPropertyDefinitions } from "../../../../cache/ontology-cache/resource-class-and-property-definitions";
 
 @JsonObject("ReadResource")
 export class ReadResource extends ReadWriteResource {
